backend/requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
Babel==2.3.4
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
entrypoints==0.3
enum34==1.1.6
flake8==3.7.9
flasgger==0.5.13
Flask==1.1.1
Flask-Cors==3.0.2
Flask-JSON==0.3.0
Flask-SQLAlchemy==2.2
gTTS==1.1.8
gTTS-token==1.1.3
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.2.2
mccabe==0.6.1
mistune==0.8.4
numpy==1.18.1
pathspec==0.7.0
pycodestyle==2.5.0
pydocstyle==5.0.2
pydub==0.18.0
pyflakes==2.1.1
pylint==2.4.4
pyrsistent==0.15.7
pytz==2019.3
PyYAML==5.3
regex==2020.2.18
requests==2.23.0
six==1.14.0
snowballstemmer==2.0.0
SQLAlchemy==1.3.13
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
zipp==3.0.0
